Splits user manager into a base class that can be reused changes base class of IdentityUser to UmbracoIdentityUser
